Directly open named pipe
Turns out it's actually possible to directly open a named pipe, so we can do away with the redirect-named-pipe-to-file-in-mod-dir work-around.

Not sure why it works now, as from earlier tests we had the impression it didn't, but seeing as this simplifies things quite a bit, I'm not complaining.

Note: this also requires changes on the `fs_mod_ros_windows` side, but those will be made in that repository.